move each sample into an individual file
