+++ samples/snippets/run_realtime_report_with_multiple_metrics_test.py
@@ -19,7 +19,19 @@ import run_report
 TEST_PROPERTY_ID = os.getenv("GA_TEST_PROPERTY_ID")
 
 
-def test_run_report_basic(capsys):
-    run_report.run_report(TEST_PROPERTY_ID)
+def test_run_report(capsys):
+    run_report.run_report_simple(TEST_PROPERTY_ID)
+    out, _ = capsys.readouterr()
+    assert "Report result" in out
+
+
+def test_run_report_with_multiple_metrics(capsys):
+    run_report.run_report_with_multiple_metrics(TEST_PROPERTY_ID)
+    out, _ = capsys.readouterr()
+    assert "Report result" in out
+
+
+def test_run_report_with_multiple_dimensions(capsys):
+    run_report.run_report_with_multiple_dimensions(TEST_PROPERTY_ID)
     out, _ = capsys.readouterr()
     assert "Report result" in out
