+++ samples/snippets/run_report_with_dimension_in_list_filter.py
@@ -15,17 +15,18 @@
 # limitations under the License.
 
 """Google Analytics Data API sample application demonstrating the usage of
-metric aggregations in a report.
+dimension and metric filters in a report.
 
-See https://developers.google.com/analytics/devguides/reporting/data/v1/rest/v1beta/properties/runReport#body.request_body.FIELDS.metric_aggregations
+See https://developers.google.com/analytics/devguides/reporting/data/v1/rest/v1beta/properties/runReport#body.request_body.FIELDS.dimension_filter
 for more information.
 """
-# [START analyticsdata_run_report_with_aggregations]
+# [START analyticsdata_run_report_with_dimension_in_list_filter]
 from google.analytics.data_v1beta import BetaAnalyticsDataClient
 from google.analytics.data_v1beta.types import DateRange
 from google.analytics.data_v1beta.types import Dimension
+from google.analytics.data_v1beta.types import Filter
+from google.analytics.data_v1beta.types import FilterExpression
 from google.analytics.data_v1beta.types import Metric
-from google.analytics.data_v1beta.types import MetricAggregation
 from google.analytics.data_v1beta.types import RunReportRequest
 
 from run_report import print_run_report_response
@@ -36,30 +37,38 @@ def run_sample():
     # TODO(developer): Replace this variable with your Google Analytics 4
     #  property ID before running the sample.
     property_id = "YOUR-GA4-PROPERTY-ID"
-    run_report_with_aggregations(property_id)
+    run_report_with_dimension_in_list_filter(property_id)
 
 
-def run_report_with_aggregations(property_id="YOUR-GA4-PROPERTY-ID"):
-    """Runs a report which includes total, maximum and minimum values for
-    each metric."""
+def run_report_with_dimension_in_list_filter(property_id="YOUR-GA4-PROPERTY-ID"):
+    """Runs a report using a dimension filter with `in_list_filter` expression.
+    The filter selects for when `eventName` is set to one of three event names
+    specified in the query."""
     client = BetaAnalyticsDataClient()
 
     request = RunReportRequest(
         property=f"properties/{property_id}",
-        dimensions=[Dimension(name="country")],
+        dimensions=[Dimension(name="eventName")],
         metrics=[Metric(name="sessions")],
-        date_ranges=[DateRange(start_date="365daysAgo", end_date="today")],
-        metric_aggregations=[
-            MetricAggregation.TOTAL,
-            MetricAggregation.MAXIMUM,
-            MetricAggregation.MINIMUM,
-        ],
+        date_ranges=[DateRange(start_date="7daysAgo", end_date="yesterday")],
+        dimension_filter=FilterExpression(
+            filter=Filter(
+                field_name="eventName",
+                in_list_filter=Filter.InListFilter(
+                    values=[
+                        "purchase",
+                        "in_app_purchase",
+                        "app_store_subscription_renew",
+                    ]
+                ),
+            )
+        ),
     )
     response = client.run_report(request)
     print_run_report_response(response)
 
 
-# [END analyticsdata_run_report_with_aggregations]
+# [END analyticsdata_run_report_with_dimension_in_list_filter]
 
 
 if __name__ == "__main__":
