+++ samples/snippets/run_report_with_multiple_dimension_filters_test.py
@@ -14,12 +14,14 @@
 
 import os
 
-import run_report
+import run_report_with_multiple_dimension_filters
 
 TEST_PROPERTY_ID = os.getenv("GA_TEST_PROPERTY_ID")
 
 
-def test_run_report_basic(capsys):
-    run_report.run_report(TEST_PROPERTY_ID)
+def test_run_report_with_multiple_dimension_filters(capsys):
+    run_report_with_multiple_dimension_filters.run_report_with_multiple_dimension_filters(
+        TEST_PROPERTY_ID
+    )
     out, _ = capsys.readouterr()
     assert "Report result" in out
