+++ samples/snippets/run_report_with_multiple_dimensions.py
@@ -14,18 +14,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Google Analytics Data API sample application demonstrating the usage of
-metric aggregations in a report.
+"""Google Analytics Data API sample application demonstrating the creation
+of a basic report.
 
-See https://developers.google.com/analytics/devguides/reporting/data/v1/rest/v1beta/properties/runReport#body.request_body.FIELDS.metric_aggregations
+See https://developers.google.com/analytics/devguides/reporting/data/v1/rest/v1beta/properties/runReport
 for more information.
 """
-# [START analyticsdata_run_report_with_aggregations]
+# [START analyticsdata_run_report_with_multiple_dimensions]
 from google.analytics.data_v1beta import BetaAnalyticsDataClient
 from google.analytics.data_v1beta.types import DateRange
 from google.analytics.data_v1beta.types import Dimension
 from google.analytics.data_v1beta.types import Metric
-from google.analytics.data_v1beta.types import MetricAggregation
 from google.analytics.data_v1beta.types import RunReportRequest
 
 from run_report import print_run_report_response
@@ -36,31 +35,28 @@ def run_sample():
     # TODO(developer): Replace this variable with your Google Analytics 4
     #  property ID before running the sample.
     property_id = "YOUR-GA4-PROPERTY-ID"
-    run_report_with_aggregations(property_id)
+    run_report_with_multiple_dimensions(property_id)
 
 
-def run_report_with_aggregations(property_id="YOUR-GA4-PROPERTY-ID"):
-    """Runs a report which includes total, maximum and minimum values for
-    each metric."""
+def run_report_with_multiple_dimensions(property_id="YOUR-GA4-PROPERTY-ID"):
+    """Runs a report of active users grouped by three dimensions."""
     client = BetaAnalyticsDataClient()
 
     request = RunReportRequest(
         property=f"properties/{property_id}",
-        dimensions=[Dimension(name="country")],
-        metrics=[Metric(name="sessions")],
-        date_ranges=[DateRange(start_date="365daysAgo", end_date="today")],
-        metric_aggregations=[
-            MetricAggregation.TOTAL,
-            MetricAggregation.MAXIMUM,
-            MetricAggregation.MINIMUM,
+        dimensions=[
+            Dimension(name="country"),
+            Dimension(name="region"),
+            Dimension(name="city"),
         ],
+        metrics=[Metric(name="activeUsers")],
+        date_ranges=[DateRange(start_date="7daysAgo", end_date="today")],
     )
     response = client.run_report(request)
     print_run_report_response(response)
 
 
-# [END analyticsdata_run_report_with_aggregations]
-
+# [END analyticsdata_run_report_with_multiple_dimensions]
 
 if __name__ == "__main__":
     run_sample()
