+++ samples/snippets/run_report_with_named_date_ranges.py
@@ -15,17 +15,16 @@
 # limitations under the License.
 
 """Google Analytics Data API sample application demonstrating the usage of
-metric aggregations in a report.
+date ranges in a report.
 
-See https://developers.google.com/analytics/devguides/reporting/data/v1/rest/v1beta/properties/runReport#body.request_body.FIELDS.metric_aggregations
+See https://developers.google.com/analytics/devguides/reporting/data/v1/rest/v1beta/properties/runReport#body.request_body.FIELDS.date_ranges
 for more information.
 """
-# [START analyticsdata_run_report_with_aggregations]
+# [START analyticsdata_run_report_with_date_ranges]
 from google.analytics.data_v1beta import BetaAnalyticsDataClient
 from google.analytics.data_v1beta.types import DateRange
 from google.analytics.data_v1beta.types import Dimension
 from google.analytics.data_v1beta.types import Metric
-from google.analytics.data_v1beta.types import MetricAggregation
 from google.analytics.data_v1beta.types import RunReportRequest
 
 from run_report import print_run_report_response
@@ -36,30 +35,29 @@ def run_sample():
     # TODO(developer): Replace this variable with your Google Analytics 4
     #  property ID before running the sample.
     property_id = "YOUR-GA4-PROPERTY-ID"
-    run_report_with_aggregations(property_id)
+    run_report_with_named_date_ranges(property_id)
 
 
-def run_report_with_aggregations(property_id="YOUR-GA4-PROPERTY-ID"):
-    """Runs a report which includes total, maximum and minimum values for
-    each metric."""
+def run_report_with_named_date_ranges(property_id="YOUR-GA4-PROPERTY-ID"):
+    """Runs a report using named date ranges."""
     client = BetaAnalyticsDataClient()
 
     request = RunReportRequest(
         property=f"properties/{property_id}",
+        date_ranges=[
+            DateRange(start_date="2020-01-01", end_date="2020-01-31", name="year_ago"),
+            DateRange(
+                start_date="2021-01-01", end_date="2021-01-31", name="current_year"
+            ),
+        ],
         dimensions=[Dimension(name="country")],
         metrics=[Metric(name="sessions")],
-        date_ranges=[DateRange(start_date="365daysAgo", end_date="today")],
-        metric_aggregations=[
-            MetricAggregation.TOTAL,
-            MetricAggregation.MAXIMUM,
-            MetricAggregation.MINIMUM,
-        ],
     )
     response = client.run_report(request)
     print_run_report_response(response)
 
 
-# [END analyticsdata_run_report_with_aggregations]
+# [END analyticsdata_run_report_with_date_ranges]
 
 
 if __name__ == "__main__":
